Encode comparison name in header link
I named a comparison "Why Engineer?" the other day, and it wouldn't load from the headers, since the correct URL was `https://coriolis.edcd.io/compare/Why%20Engineer%3F` but the header linked to `https://coriolis.edcd.io/compare/Why%20Engineer?`. This just encodeURIComponent()'s the name to avoid that.

NOTE: I wasn't able to spin up a dev environment to test this change, but it appeared straightforward enough that I figured it was quicker to send a PR than it was to open an issue.